Пересмотрена файловая структура проекта и добалены задачи
