+++ .env
@@ -1,5 +1,6 @@
-Port=:5432"
-User="postgres"
-Password="postgres"
-Hotels_Data="Booking Data"
-Bookings_data="Hotel Data"
+Host=localhost
+Port=5432
+User=postgres
+Password=postgres
+HotelDB=Hotels Data
+BookingDB=Booking Data
