--- consts/value.go
@@ -1,3 +1,0 @@
-package consts
-
-const MaxInt = ^(0)
