--- db/hotels_data/GET/get_hotel_by_rating.go
@@ -1,17 +1,0 @@
-package hotels_data
-
-import (
-	"github.com/maxturyev/booking-system-project/models"
-	"gorm.io/gorm"
-)
-
-func GetHotelByRating(db *gorm.DB, stars ...int) []models.Hotel {
-	var hotels []models.Hotel
-
-	result := db.Where("rating IN ?", stars).Find(hotels)
-	if result.Error != nil {
-		panic("Error")
-	}
-
-	return hotels
-}
