--- db/hotels_data/POST/change_hotel_count_of_room.go
@@ -1,1 +1,0 @@
-package hotels_data
