--- db/hotels_data/POST/change_hotel_price.go
@@ -1,1 +1,0 @@
-package hotels_data
