+++ src/hotel-svc/configs/config.yml
@@ -1,0 +1,8 @@
+server:
+  host: 127.0.0.1
+  port: 9090
+  timeout:
+    server: 30
+    read: 15
+    write: 10
+    idle: 5
