+++ src/hotel-svc/databases/databases.go
@@ -1,0 +1,6 @@
+package databases
+
+import "gorm.io/gorm"
+
+// Database shares global database instance
+var Database *gorm.DB
